Crate OispaHallaAnticheat

Vilpinestopalvelin peliin 2048 pohjautuvalle oispahalla:lle, pohjautuu kirjastoon twothousand-forty-eight.

Lataaminen

Tämä repo käyttää git -submoduuleja, jotka pitää ladata ennen projektin koontia.

Voit ladata ne automaattisesti repon kanssa samaan aikaan käyttämällä komentoa git clone --recurse-submodules git@github.com:hallabois/OispaHallaAnticheat.git tai suorittamalla komennon git pull --recurse-submodules aiemmin kloonatussa repossa.

Submoduulit voi päivittää myöhemmin komennolla git submodule update --remote --merge.

Koonti ja suorittaminen

Käyttää rustia, asenna se aluksi: https://www.rust-lang.org/tools/install

Huom: palvelin-ominaisuus vaatii vieläpä rustin nightly-version, josta voit asentaa tuoreen version komennoilla rustup update && rustup default nightly

HTTPS-tuki

aja palvelin komennolla TLS_CERT="/etc/letsencrypt/live/hac.hallacoin.ml/fullchain.pem" TLS_KEY="/etc/letsencrypt/live/hac.hallacoin.ml/privkey.pem" ./target/release/g2048engine --server

Projektin Tiedostorakenne

Lue ARCHITECTURE.md.

API:n Käyttö

/HAC/validate

https://hac.oispahalla.com:8000/HAC/validate/[leveys]x[korkeus]S[peli]
---------------------------------------------
{
    "run_hash":"724...",   Pelin uniikki tunniste. Älä käytä, jos peli ei ole validi
    "board_w":4,           Tulkittu pelilaudan leveys
    "board_h":4,           Tulkittu pelilaudan korkeus
    "valid":true,          Onko syöte validi peli
    "score":584,           Mitä pelin pisteiden tulisi olla*
    "score_margin":4,      Viimeisen vuoron aiheuttama mahdollinen pisteiden liikkumisvara*
    "breaks":0,            Kuinka monta kurinpalautusta pelin aikana käytettiin*
    "length":85            Pelin pituus siirroissa
} 
*palauttaa nollan, jos peli ei ole validi

e.g. https://hac.oispahalla.com:8000/HAC/validate/liianpitkätähän tai minimi-pituinen peli https://hac.oispahalla.com:8000/HAC/validate/4x4S0.0.0.0.0.0.0.0.0.0.0.0.0.0.2.2+2,1.2;1

/HAC/get_config

https://hac.oispahalla.com:8000/HAC/get_config
---------------------------------------------
{
  "platform": "x86_64-unknown-linux-gnu",                          millä alustalla kyseinen instanssi pyörii
  "version": "febc9c91bd18d4be6b4989e3d24898c9bb12ca84",           mikä oli viimeisin julkaistu versio tai git-commit ennen kasaamista
  "rust_version": "rustc 1.57.0-nightly (e1e9319d9 2021-10-14)",   millä rustin versiolla projekti on kasattu
  "request_count": 0                                               kuinka monta kertaa /validate komentoa on kutsuttu
}

Modules

server 🔒

Structs

Args 🔒

An anticheat solution for the 2048 fork OispaHalla

Constants

DEBUG_INFO 🔒

Näyttää lisää virheenkorjaustietoja tarvittaessa

Functions

game 🔒

main 🔒